documentation: Re-defined "Build Appliance" to mean the file
Previous renditions of the term "Build Appliance" referred to the
VMware system.  However, it was determined that the term
"Build Appliance" should really refer to the actual image that
can be booted and run on the VMware applications.  I have updated
the areas affected by this change.  The change assumes the name of
the built or downloaded file you boot will be named
"build-appliance-image."

Fixes [YOCTO_#2656]

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>