documentation/poky-ref-manual: Removed artifacts for dev chapter
Removing the Development chapter resulted in some links being
re-defined to go to the YP Development manual.  Also, the
poky-ref-manual.xml file had to be altered to not include the
chapter.  Any sort of residual cleanup needed as a result of
getting rid of this chapter is handled in this commit.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>