documentation: Changed link anchor for 'resourses' in QS
The string used for the link anchor for the "What You Need and
How You Get It" section in the YP QS used "resources".  This
string is actually used as the chapter id for the resources
chapter in the YP ref manual.  I changed the string in the QS
from "resources" to "yp-resourses" to make it unique.  There was
one instance in the YP ref manual where a cross-reference to the
section occurred.  I fixed that as well.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>