documentation/poky.ent: updated YOCTO_DOCS_RELEASE variable to 1.3
We should not use "latest" any longer for this variable.  The
string is fine in URLs that might appear on a website but the actual
variable should be set to the release number.  This way when we
go to release the version of the docs they do not have to be reset
with a post-release commit to change the variable from "lastest" to
"1.3".

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>