documentation: Moved the Eclipse chapter to the YP Dev Manual.
The detailed chapter that describes how to install and configure
the Eclipse Yocto Plug-in has been moved to become a subsection of
the "Workflow" chapter in the YP Development Manual.  This commit
accomplished the bulk move, edits to integrate the information,
and fixes to all cross-references.  Moving the chapter creates a
sectioning issue in the YP Development manual that results in
subsections that are six levels deep.  Not ideal.  Will look at
fixing these with a subsequent commit.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>