documentation/poky-ref-manual: Removed included BSP Guide
With the decision to create a single HTML file that has all
individual YP manuals, the need for including the BSP guide as
part of the YP Reference Manual goes away.  I have removed this
guide as a chapter, removed the mention of it in the introduction,
and fixed the two links in the FAQ chapter to reference the external
manual.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>