documentation: tag change adt and kernel manuals
I changed a couple of id strings so that they are not duplicates.
I did this to eliminate a make error while experimenting with creation
of the single HTML YP document comprised of the individual docs.
Changing the id strings has no consequence on the individual Application
Developer's Guide and Kernel Manual.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>