documentation: Style sheet naming changes
I had to rename individual style.css sheets to have book-specific
names.  So style.css for the ADT manual becomes adt-style.css and
so forth.  Turns out, this isn't really necessary for the solution
to create the mega-manual but I am leaving the changes as a
separate commit because it makes sense.  It causes no disruption
to making the six YP manuals.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>