documentation: Fixes to note for mega-manual processing
the notes that say "Due to production processes..." at the start of
each manual contain links to manuals.  the links needed to be formatted
such that the entire <ulink> code and manual name appeared on the same
line so that the sed script can effectively remove the link in the
resulting mega-manual.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>