documentation/dev-manual/dev-manual-kernel-appendix.xml: Fix PARALLEL_MAKE
Every place in the YP documentation suggests setting BB_NUMBER_THREADS
and PARALLEL_MAKE to twice the number of cores as the host machine.
This fixes a single instance in the YP dev manual where it was
suggested that PARELLEL_MAKE be set at 1.5x times the number of
cores.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>