documentation/mega-manual: New mega-manual directory
This new directory contains the files for the mega-manual HTML
document.  Mega-manual is a compilation of all the existing YP
manuals and guides.  It is a single HTML file that simply lists
each individual HTML document for the existing YP manuals and
guides.  The figures directory contains all the figures (duplicated)
from the individual manuals and guides.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>