documentation: Fixed links to BitBake User Manual
Robert P. J. Day pointed out a link in the development manual that
was for the BitBake User Manual.  The link resolved to an old,
non-maintained version of BitBake documentation.  According to
Richard Purdie, the BitBake docs referenced should be the ones
that ship in the bitbake/doc/manual directory of poky.  The
YP docs had been using a variable named BITBAKE_DOCS_URL that
also resolved to the antiquated user manual site.  Consequently,
there were many links that needed fixed across both the YP
Development Manual and the YP Reference Manual.  Each of these
references now points in general to the bitbake/doc/manual
directory in poky for more information.

Reported-by: Robert P. J. Day <rpjday@crashcourse.ca>
Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>