documentation/yocto-project-qs/yocto-project-qs.xml: Adjust GUI component feature listing
List frameworks we provide that are likely to be used by people these
days - Pimlico has been removed and GuPNP and Matchbox don't really
belong. Also slightly modify the sentence following so that it is clear
that these are optional.

In addition to Paul's patch, I cleaned up one of his sentences with
some simple grammar edits.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>