documentation: poky-ref-manual, mega-manual, Makefile - new variables
Added variable descriptions for BB_DISKMON_DIRS and
BB_DISKMON_WARNINTERVAL.  These went in the reference
manual in the glossary section.  First draft for both
descriptions.

Also, updated the Makefile to include two new figures used
in the Build History section.  The figures needed to be added
to the TARFILE variable so that the .PNG files are included
when publishing both the poky-ref-manual and the mega-manual.

Added two figures to the figures directories of the poky-ref-manual
and the mega-manual.  These figures are part of the new Build
History section.  Anytime you add a figure to any manual you
also have to store it in the figures directory of the mega-manual.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>