documentation: adt-manual - autotools-based projects
Fixes [YOCTO #2645]

In the section that describes how to build a toolchain using
bitbake meta-toolchain, I updated the note about the MACHINE
and SDKMACHINE variables.  These needed to be pointed out
to the user as important variables for correct settings
before generating the toolchain.

I added a new section for the autotools-based programs that
now includes a simple "Hello World" example.  This section
precedes the section that describes how to pass host options
to configure.sh.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>