documentation: dev-manual - Updated collaboration section
Fixes [YOCTO #3274]

I have added a new area at the bottom of the current
"Using the Yocto Project in a Team Environment" section.
This section is the spot for addressing what Richard Purdie
coined as "How to use YP in a work group."  For now, the
changes are based on the information that Richard placed in
the Bugzilla entry.  Presumably we can build on this information.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>