Documentation: dev-manual, bsp-guide, poky-ref-manual - bblayer.conf updates
Fixes [YOCTO #3176]

Several places throughout the YP manual set examples are given
for the bblayer.conf file.  For release 1.3, a bug existed that
surfaced when a user removed either the meta or meta-yocto
layer from this configuration file.  The bug has been fixed and
the ripple effect through the YP documentation set affected
several manuals.

Fixes include updating the code samples that show bblayer.conf
so that they now include the BBLAYERS_NON_REMOVABLE variable,
a new glossary entry for the variable, and an explanation of
both BBLAYER and BBLAYER_NON_REMOVABLE in the section that
talks about the bblayer.conf file.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>