kernel-dev: Updates to get the new kernel-dev manual to make
These changes were to the intro chapter to introduce the DocBook
form, the Makefile to introduce the new manual, the figure used
for the title had to be updated, the root manual .XML file needed
changed, and the style sheet.

The manual now makes with just the raw text for Darren's intro
section.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>