kernel-dev: Created super section for Branches.
The sections for "Machine Branches" and "Feature Branches" really
should be encapsulated within a section on branching.  Much of
the text in the "Machine Branches" section presents the rationale
behind both types of branches.  I have created a larger section
to contain these and will edit the text to fit.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>