kernel-dev: Edits to "Creating and Preparing a Layer"
Changed the name of this section and took out the bit about
the manual being written with the assumption of a custom
layer being created.  The manual does not particularly call
this out with any examples so it is not needed.

I pulled the copy of the original text also.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>