dev-manual: Scrub for correct YP manual cross-references.
I went through and got rid of all references to the old
YP Kernel Architecture and Use Manual and replaced with
appropriate references to the new YP Linux Kernel
Development Manual.

I also laced in appropriate references into various sections
of the new YP Profiling and Tracing Manual.

Also, updated the list of resource manuals to reflect the
new manual set (addition of kernel-dev and profile-dev and
removal of kernel-manual).

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>