kernel-dev: Added "Kernel Architecture" section.
Moved the "Kernel Architecture" section from the YP Kernel
Architecture and Use Manual to this manual.  The section
included the kernel-architecture-overview.png figure.  So,
I added that PNG file to the "figures" folder.  Finally, I
had to also add the PNG file to the Makefile tarfile list
for kernel-dev.  Note that because the figure was part of
the old YP Kernel Architecture and Use Manual, I did not have
to add the figure to the mega-manual tarfile list.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>