kernel-dev: Added "Maintenance" appendix.
New appendix required adding the kernel-dev-maint-appx.xml
file.  The contents are sections 3.2 and 3.3 from the old
YP Kernel Architecture and Use Manual, which is "Tree
Construction" and "Build Strategy."  According to Darren Hart,
this information should be retained for the few people on
the planet that might ever need it.  In addition to adding
the new file, I had to add the bits to the kernel-dev.xml
manual to support including the appendix when making the
manual.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>