kernel-dev: Added "-dirty String" section.
The "-dirty" String section from the old YP Kernel Architecture
and Use Manual was moved to section 2.3 of the manual.  It needs
to find the right area to be put in (AR for Darren).  I added the
AR text to the top of the section so it would not fall through
the cracks.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>