dev-manual: Updates to expose cross-toolchain recipes.
Added a bulleted list to the "Terms" section to help define
some toolchain recipes such as gcc-cross-initial,
gcc-cross-intermediate, and so forth.

Added a cross-reference link to the term
"Cross-Development Toolchain" in the the "Working Using
Stand-alone Cross-development Toolchains" section.
This gives the reader a way to see more on the toolchain
recipes.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>