template:  Changed the font color for footers and headers
Fixes YOCTO #3728

The poky-db-pdf.xsl file has code that controls the font color
for the headers, footers, and separating lines for each when
the PDF version of a manual is made.  I changed the values
from #cccccc to #999999 to darken these up so they can be
seen a little better.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>