dev-manual: Added notes about creating your own distro
Added a couple of notes to point out the ease of creating your
own distro and not just relying on a Poky distribution.  One
note is at the very beginning of the manual and mentions
Angstrom as an example of a distro based on the YP.  The
other note is at the beginning of the second chapter.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>