ref-manual: Applied Paul Eggleton's buildhistory patch 2 of 5.
Additional information added to the end of the
"Build History Package Information" section."

Buildhistory now collects information on SRCREV values for recipes
fetched from a version control system e.g. Git; additionally a
buildhistory-collect-srcrevs tool is provided to gather this and
convert it to a format suitable for inclusion in global configuration.
Add information on these new features to the appropriate section.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>