dev-manual, bsp-guide: Updates for init-ifupdown change.
Release 1.4 requires the user to create an append file to the
new recipe init-ifupdown if they want to create their own
custom /etc/network/interfaces file.  Previously, they needed
to append the netbase recipe.

Changes needed in the migration section of the ref-manual as
well as the exmaple used in the bsp-guide.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>