kernel-dev: Updates to "Creating the Append File" section.
Pointed out by Robert P. J. Day that I was using the wrong
FILESEXTRAPATHS example for an append file.  I failed to include
the "_prepend" part of the variable.  Also, the wording based
on the example made some assumptions and was not entirely
accurate.  The changes I put in made the section very specific.

Reported-by: Robert P. J. Day <rpjday@crashcourse.ca>
Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>