yocto-project-qs: Update to Linux Distribution Requirements
Fixes [YOCTO #4753]

The section that tells the user the basic Linux requirement has
been updated to mention the Git, tar, and Python minimum
requirements.  There is a new section in the reference manual's
introduction area that describes how to create the buildtools
tarball if these requirements are not met.  The changes I made
in the Quick Start now summarize the requirements and provide
a link to that new section in the YP Reference Manual.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>