documentation: Removed deprecated kernel-manual.
Fixes [YOCTO #4810]

I removed this manual from the tip of the repo.  We carried a symlink
to it for the 1.4 and 1.4.1 release.  Now it is time to get rid of it.
Removing this manual should get rid of the folder in the poky/documentation
folder for future clones of that repo.

This bug was entered because someone tried to make the manual for YP
1.5.  It is debatable as to whether this is a fix or not.  You might
argue that the bug is not even a bug.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>