dev-manual, ref-manual, yocto-project-qs: Debian support added
Fixes [YOCTO #4912]

Throughout the manual set, various areas talk about the Linux
distributions we support and the required packages among other
items that are tied to a distro.  Debian had been missing except
from the list in the ref-manual that shows the releases we test
against.

I have attended to all areas where Debian needs to be mentioned
now.

Reported-by: Laszlo Papp <lpapp@kde.org>
Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>