ref-manual, Makefile: New section on BitBake for expanded build
Fixes [YOCTO #2808]

Created a new section called "BitBake" that will provide an
expanded discussion of the basic YP build process.  The
"BitBake" section has a new sub-section called "Source Fetching."

Part of the new section is a new figure named source-fetching.png.
The figure must reside in the "figures" directory in both the
ref-manual and the mega-manual.  It must also be listed for each
of these manuals in the TARBALL statement in the Makefile.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>