ref-manual: Created new "Closer Look" chapter
Fixes [YOCTO #2808]

I extracted the section that takes a closer look at the YP
development process into its own chapter.  Feedback during the
review indicated that this information should not be buried as
it was in a section but rather pulled higher out for visibility.

So, The changes create a new chapter three that is dedicated to
this topic.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>