ref-manual: Added information for OEROOT and build environment
Fixes [YOCTO #4980]

Added a new glossary entry for OEROOT.

Added significant information to the local.conf and bblayers.conf
reference sections to describe how the root build directory is
derived.

Also, some unrelated formatting to the DL_DIR variable description
was mixed in here because I forgot to commit that separatey and it
lives in the same variables.xml file.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>