ref-manual, mega-manual: Updates to "Image Generation"
Fixes [YOCTO #2808]

Updated the figure and placed it in the folders area for both
the ref-manual and mega-manual.  I had to add the
DEPLOY_DIR_IMAGE variable and create a new machine-specific
directory below deploy.  The text had to be adjusted to reflect
this change as well.

Also, created a new variable entry for DEPLOY_DIR_IMAGE for
the glossary of the ref-manual.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>