ref-manual, yocto-project-qs: Explanations for Git tarballs
Fixes [YOCTO #5035]

By default, the DL_DIR does not get files from Git repositories
that are suitable for mirroring.  There is a work-around by
using the BB_GENERATE_MIRROR_TARBALLS variable, which, when
set, causes the build system to generate tarballs for the
Git repositories and place them in the DL_DIR.

To address this, I added a new variable description for the
BB_GENERATE_MIRROR_TARBALLS varible, updated the DL_DIR
variable to indicate that "out-of-the-box" the YP puts suitable
mirror files into the DL_DIR for everthing except Git
repositories.

Some other sections were updated where discussion revolved around
pulling down source files.  The QS had a spot where the
BB_GENERATE_MIRROR_TARBALLS variable was actually used in an
example.  And, the expanded discussion on the BitBake process
had a couple spots that deserved a mention.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>