dev-manual: Patch applied to runtime package management.
Trevor Woerner's patch as submitted with no alterations.
This patch rewrites the entire section on setting up
runtime package management.  The section will need to be edited
but that will be a separate exercise.  This commit captures
the raw patch.

Note: one external link in the doc set broken after application
of the patch because a "section id" tag was changed by the
patch.  I am leaving that fix to a separate commit.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>