ref-manual, dev-manual: Edits to runtime package management section.
These changes modify the patch sent by Trevor that essentially
re-wrote this section.  My edits were for consistency only and
style of the book.  No technical information or flow was altered.

I did rename the section to be active to match the other package
related sections.  This caused the link in the ref-manual in the
classes chapter to have to be updated.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>