documentation: Changed Note for most recent manual
Fixes [YOCTO #5368]

Changes partially address this issue, which is the removal of
tarball installations for poky.  The notes at the beginning of
the YP manuals suggested that the tarball version of a manual
might lag the version found on the website.  Because we are
discouraging installing poky/documentation from a tarball now,
I have re-written the note to be generic and suggest simply that
for the most recent version of the manual associated with the
release, see the manual on the website.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>