dev-manual: Edits to "Yocto Project Source Repositories" section
Fixes [YOCTO #5368]

This is a partial fix to the removing tarball setup method issue.
I have modified the "Yocto Project source Repositories" section to
de-emphasize the tarball installation method for poky and meta-intel.
We cannot remove these obviously because the download page of the
website gives the use access to tarballs.  Also, the Index of
Releases basically does the same thing.  So, we can't ignore the
fact that they are there. So, I added a note basically saying we
recommend the Git method for creation of poky and meta-intel on the
development system.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>