ref-manual: updated the list of where to get YP release
Partial fix for [YOCTO #5368]

The list that summarized where to get YP release did not include
the source repositories.  It basically just listed tarball
areas.  Now that cloning the Git repo of poky is the preferred
method for obtaining YP, I made sure that this area was the
first spot listed and called out as the preferred method.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>