dev-manual: Edits to the toaster section
Partial fix for [YOCTO #4414]

Got some feedback on the section and added a step that the user
needs to checkout the dora-toaster branch after clonine poky.
Also, that the Django version is specific and not the listed
version +.  Finally, a bit of wording to note that the GUI is
temporary for this release.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>