dev-manual: A few tweaks to Ch2 of dev-manual
scott (or anyone else) is welcome to use any or all of this, or
tweak to taste. i have a few other concerns with ch 2 but i'll read it
more carefully to make sure i'm reading it correctly. - rpjday

I implemented all but the addition of MIPS64 as it is not tested
using the autobuilder yet.  - scottrif

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>