dev-manual: Fixed broken structure.
I discovered that the dev-manual was not making correctly.
Probably resulted in trying to merge in some of Robert P. J. Day's
comments.  I have made the tag changes to fix this.  Also,
I added back in the two methods for setting up meta-intel
as I am not making the tarball-exclusion changes to the dora
branch.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>