ref-manual: Improved on package* class.
Previously, we were documenting the "package*" class and lumping
the "package_deb", "package_rpm", and "package_ipk" classes in
that entry.  Really, we need to break out the "package" class on
its own and create entries for the sub-classes that were being
bundled in there.  Additionally, we needed to document the
"package_tar" class.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>