ref-manual: Edits to icecc class and removal of three variables.
Fixed the links in the icecc class that linked into ICECC_CC,
ICECC_CXX, and ICECC_VERSION so they don't link anymore but rather
give a very brief explanation of what the variable does.

Removed the above three variables from the variable glossary.
These are not BitBake variables and we should not document them.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>