dev-manual: Edits to the "Post-Installation Scripts" section.
There was some confusion to this section as to just when post-install
scripts were being executed.  I looked it over and the section really
is focusing more on the situations when you are creating a read-only
root filesystem.  So I went through and did a small amount of
editing to just clean up some loose text.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>