dev-manual: Edits to "Packaging" section.
Applied some self-edits to this section that is part of the
"Writing a New Recipe" section.  There was no feedback from the
second review but I noticed a few issues.  Also, I added that
reference on how to add a post-installation script from the
"Enabling System Services" section as noted by Paul Eggleton.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>