ref-manual: Edits to "Understanding What the Build History Contains" section.
The opening sentence here said that the build history is
kept in the $TMPDIR/buildhistory directory in the Build Directory
as defined by the BUILDHISTORY_DIR variable.  Well, the variable
claims it is kept in $TOPDIR/buildhistory.  I went with the
variable as the correct source.  So, I changed the intro
sentence here.

Reported-by: Laszlo Papp <lpapp@kde.org>
Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>