ref-manual: New variables added and some edits as well
Fixes [YOCTO #5931]

Added descriptions for the following variables in the glossary:

* MODULE_IMAGE_BASE_NAME
* KERNEL_IMAGE_BASE_NAME
* PKGE
* PKGV
* PKGR
* PKG

The first two are for the bug.  The remainder needed to be added
as they were referenced from the first two (except for PKG, which
I was told needed to be added anyway.  I also fixed a couple formatting
issues in there for some other variables.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>