adt-manual: Restructured opening organization.
The start of this manual was odd.  I created a new "Introduction"
chapter and inserted it into the build file (adt-manual.xml).
This new chapter introduces the manual.  That left the remainder
of the original first chapter to be able to focus on what it
really needed to talk about and that is the ADT.  It is a better
organization.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>