dev-manual: Added yocto-layer script note and updated kernel version
I added a bit about the layer creation in the patching the kernel
section.  The example does it by hand and now you could obviously
use the yocto-layer create script.  I made a note of saying the
example does it by hand.

Also, in the menuconfig example.  I updated the kernel version from
3.4 to 3.14.  It was very old.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>