ref-manual, mega-manual: Added manifest file information
Fixes [YOCTO #5410]

I updated the figure that shows what do_rootfs does to create
the image.  The new addition is the creation of the .manifest file
that lists out package information in the image.  Supporting text
was also added in the section.

The updated figure was needed in both the ref-manual and
mega-manual figures folders.

Also, part of the fix included a new variable description for the
IMAGE_MANIFEST variable.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>