ref-manual, yocto-project-qs: Notes and infor for buildtools issues
Fixes [YOCTO #4866]

Updated the main area where it talks about needing the right
versions of Git, tar, and Python.  You can get this set up using a
couple methods.  I documented both.

The changes in the FAQ and in the QS are basically notes that point
to that section for more information.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>